phonebook-api/pyproject.toml
[tool.poetry]
name = "phonebook-api"
version = "0.1.0"
description = ""
authors = ["mogendi <mogendi.mongare@gmail.com>"]
readme = "README.md"
packages = [{include = "phonebook_api"}]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.5"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
